layout/base/tests/chrome/test_bug812817.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=812817
-->
<window title="Mozilla Bug 812817"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="doTest()">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<panel id="panel" width="200" height="200" onpopupshown="continueTest()">
</panel>
<body xmlns="http://www.w3.org/1999/xhtml">
<a href="https://bugzilla.mozilla.org/show_bug.cgi?id=812817"
   target="_blank">Mozilla Bug 812817</a>
</body>
  <!-- test code goes here -->
<script type="application/javascript">
<![CDATA[
SimpleTest.waitForExplicitFinish();

var panel = document.getElementById('panel');
function doTest() {
  panel.openPopup(null, '', 500, 500, false, false, null);
}

function continueTest() {
  panel.style.background = "url(blue-32x32.png)";
  setTimeout(function() {
    ok(true, "Didn't crash");
    SimpleTest.finish();
  }, 50);
}

]]></script>
</window>
